refactor: solve circular reference of Table
